Stop using the inl pattern in automotive
It was an example for others to refer to, but isn't very useful for
these specific classes.  It's most appropriate in the framework classes,
where is still has some example uses.
