+++ drake/automotive/BUILD
@@ -37,13 +37,13 @@ cc_library(
 cc_library(
     name = "idm_with_trajectory_agent",
     srcs = ["idm_with_trajectory_agent.cc"],
-    hdrs = ["idm_with_trajectory_agent.h", "idm_with_trajectory_agent-inl.h"],
+    hdrs = ["idm_with_trajectory_agent.h"],
     deps = DEPS,
     linkstatic = 1)
 cc_library(
     name = "simple_car",
     srcs = ["simple_car.cc"],
-    hdrs = ["simple_car.h", "simple_car-inl.h"],
+    hdrs = ["simple_car.h"],
     deps = DEPS,
     linkstatic = 1)
 cc_library(
