+++ drake/automotive/simple_car.h
@@ -12,9 +12,6 @@ namespace automotive {
 /// SimpleCar -- model an idealized response to driving commands, neglecting
 /// all physics.
 ///
-/// This class uses Drake's `-inl.h` pattern.  When seeing linker errors from
-/// this class, please refer to http://drake.mit.edu/cxx_inl.html.
-///
 /// configuration:
 /// * see lcmt_SimpleCarConfig_t
 ///
@@ -43,8 +40,6 @@ namespace automotive {
 ///
 /// They are already available to link against in libdrakeAutomotive.
 ///
-/// To use other unusual ScalarType substitutions,
-/// see http://drake.mit.edu/cxx_inl.html.
 /// @ingroup automotive_systems
 template <typename T>
 class SimpleCar : public systems::LeafSystem<T> {
